# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "This is the API documentation generated by Swagger for my Laravel project",
        "contact": {
            "email": "youremail@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/user/profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get user's profile",
                "operationId": "11670214cec0da99c9ebdb360863a980",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update user's profile",
                "operationId": "e282ccab16af87b9c989726c3203be39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "name@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "01234567890"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Cairo"
                                    },
                                    "image": {
                                        "description": "Profile image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Delete user's profile",
                "operationId": "d4e039659b9d0c31fdf001ff3bc88560",
                "responses": {
                    "200": {
                        "description": "Profile deleted successfully"
                    },
                    "404": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/profile/change-password": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Change user's password",
                "operationId": "735714428db5501ffc098d843d87fd45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "currentPassword123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully"
                    },
                    "400": {
                        "description": "Current password is incorrect"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "User Profile",
            "description": "User Profile"
        }
    ]
}